Organize files for ball-in-maze sample and separate code into different files.
