+++ ball-in-maze/src/header/globals.h
@@ -1,0 +1,21 @@
+#pragma once
+
+//Panda3D includes
+#include "pandaFramework.h"
+#include "collisionHandlerQueue.h"
+
+namespace Globals {
+	extern PandaFramework framework;
+	extern WindowFramework* window;
+	extern LVector3 ballV;
+	extern LVector3 accelV;
+	extern PT(CollisionHandlerQueue) collision_handler;
+	extern NodePath maze;
+	extern NodePath ball;
+	extern NodePath ball_root;
+
+	//Some constants for the program
+	extern int ACCELERATION;
+	extern int MAX_SPEED;
+	extern int MAX_SPEED_SQ;
+}
