+++ ball-in-maze/src/main.cpp
@@ -8,108 +8,13 @@
 #include "ambientLight.h"
 #include "directionalLight.h"
 
+#include "header/functions.h"
+
 std::string samplePath = "./";
-WindowFramework* window;
-PT(CollisionHandlerQueue) collision_handler;
-NodePath maze = NodePath("maze");
-NodePath ball = NodePath("ball");
-NodePath ball_root = NodePath("ball_root");
-LVector3 ballV;
-LVector3 accelV;
 std::vector<NodePath> lose_triggers;
 
-//Some constants for the program
-int ACCELERATION = 70;							//Acceleration in ft/sec/sec
-int MAX_SPEED = 5;								//Max speed in ft/sec
-int MAX_SPEED_SQ = 25;							//MAX_SPEED squared
-
-//Exit function
-void panda_exit(const Event* theEvent, void* data) {
-	exit(0);
-}
-
-void wall_collide_handler(PT(CollisionEntry) entry) {
-	LVector3 norm = entry->get_surface_normal(window->get_render()) * -1;
-	double current_speed = ballV.length();
-	LVector3 inVec = ballV / current_speed;
-	double vel_angle = norm.dot(inVec);
-	LVector3 hit_dir = entry->get_surface_point(window->get_render()) - ball_root.get_pos();
-	hit_dir.normalize();
-	double hit_angle = norm.dot(hit_dir);
-
-	if (vel_angle > 0 && hit_angle > 0.995) {
-		//Standard reflection equation
-		LPoint3 reflectVec = (norm * norm.dot(inVec * -1) * 2) + inVec;
-
-		ballV = reflectVec * (current_speed * (((1 - vel_angle) * 0.5) + 0.5));
-
-		LPoint3 disp = entry->get_surface_point(window->get_render()) - entry->get_interior_point(window->get_render());
-		LVector3 newPos = ball_root.get_pos() + disp;
-		ball_root.set_pos(newPos);
-	}
-	printf("Wall collide handler called!\n");
-}
-
-void ground_collide_handler(PT(CollisionEntry) entry) {
-	double newZ = entry->get_surface_point(window->get_render()).get_z();
-	ball_root.set_z(newZ + 0.4);
-
-	LVector3 norm = entry->get_surface_normal(window->get_render());
-	LVector3 accelSide = norm.cross(LVector3::up());
-
-	accelV = norm.cross(accelSide);
-	printf("Ground collide handler called!\n");
-}
-
-//Roll task
-AsyncTask::DoneStatus roll_func(GenericAsyncTask* task, void* mouseWatcherNode) {
-	double dt = ClockObject::get_global_clock()->get_dt();
-	PT(MouseWatcher) mouseWatcher = DCAST(MouseWatcher, (PandaNode*)mouseWatcherNode);
-	if (dt > 0.2) {
-		return AsyncTask::DS_cont;
-	}
-
-	if (collision_handler->get_num_entries() > 0) {
-		collision_handler->sort_entries();
-		for (size_t i = 0; i < collision_handler->get_num_entries(); i++) {
-			PT(CollisionEntry) entry = collision_handler->get_entry(i);
-			std::string name = entry->get_into_node()->get_name();
-			std::cout << name << std::endl;
-			if (name == "wall_collide") {
-				wall_collide_handler(entry);
-			} else if (name == "ground_collide") {
-				ground_collide_handler(entry);
-			} else if (name == "loseTrigger") {
-				//lose_game(entry);
-			}
-		}
-	} else {
-		printf("No entries!\n");
-	}
-
-	if (mouseWatcher->has_mouse()) {
-		LPoint2 mpos = mouseWatcher->get_mouse(); //Get the mouse position
-		maze.set_p(mpos.get_y() * -10);
-		maze.set_r(mpos.get_x() * 10);
-	}
-
-	ballV += accelV * dt * ACCELERATION;
-	if (ballV.length_squared() > MAX_SPEED_SQ) {
-		ballV.normalize();
-		ballV *= MAX_SPEED;
-	}
-	ball_root.set_pos(ball_root.get_pos() + (ballV * dt));		//Update the position based on the velocity
-	LRotation prevRot(ball.get_quat());
-	LVector3 axis = LVector3::up().cross(ballV);
-	LRotation newRot(axis, 45.5 * dt * ballV.length());
-	ball.set_quat(prevRot * newRot);
-
-	std::cout << ball_root.get_pos() << std::endl;
-
-	return AsyncTask::DS_cont;
-}
-
 int main(int argc, char* argv[]) {
+	using namespace Globals;
 
 	//Check if any arguments was given at startup
 	if (argc > 1) {
@@ -119,7 +24,6 @@ int main(int argc, char* argv[]) {
 	}
 
 	//Open framework
-	PandaFramework framework;
 	framework.open_framework(argc, argv);
 
 	//Set the window title and open new window
@@ -183,7 +87,6 @@ int main(int argc, char* argv[]) {
 
 	//Collision
 	CollisionTraverser collision_traverser;
-	collision_handler = new CollisionHandlerQueue();
 	collision_traverser.add_collider(ball_sphere, collision_handler);
 	collision_traverser.add_collider(ball_ground_col_NP, collision_handler);
 	collision_traverser.show_collisions(window->get_render());
@@ -209,8 +112,6 @@ int main(int argc, char* argv[]) {
 
 	//More initialization
 	ball_root.set_pos(maze.find("**/start").get_pos());
-	ballV.set(0, 0, 0);
-	accelV.set(0, 0, 0);
 	PT(GenericAsyncTask) roll_task = new GenericAsyncTask("roll_task", roll_func, (void*)window->get_mouse().node());
 	AsyncTaskManager::get_global_ptr()->add(roll_task);
